Add precise information about driver
CUDA 5.0 introduced a driver bug (mainly ECC false negative
errors) between the 5.0 early access/release candidate and the
community release. Due to that, we simply recommend to update
to the latest 5.5 or soon to be 6.0 drivers.

Remember: you can always run an *older* cuda toolkit with a
newer driver (but not vice versa).