Make ionization folder
Change of plans. There will be no ionization plugin anymore. It is crucial
when ionization happens, so it will be put directly into the PIC cycle.
This commit contains uncommited changes to old non-working plugins to
touch the attributes of particles and a new folder called "ionization"
with an "ionization.hpp".