Last changes for ionization plugins - then abort
These most recent changes to the non-functional plugins for "touching the
particles" will only be committed to be used for writing a guide as to how
plugins are witten. Then they will be removed because ionization will not
be implemented in the form of plugins anymore but directly into the code.