Add electron write function
When electrons are created they are cloned from their parent ion and their
attributes are copied. Some attributes are deselected in the copying process
to be set hard later. Those are the multiMask, chargeState and momentum.
Copying the first one would be slow, copying the others would just be wrong.