TBG: Add Defaults via Env for -s|-t
This feature shortens the tbg call dramatically without harming
its generality.

The -s | --submit option is the batch system and unique for each
machine. That is a perfect candidate for a picongpu.profile or
a general .profile setting (as a default).

The -t | --tpl file is useful if you have various queues or debug
templates. 99% of the users always submit to the same queue and
would need a default, too. Developers or users with varying queues
can still use -t to overwrite the default.