HDF5 WriteField: Fix Wrong Doc String and Index
the index names in `hdf5/writer/Field.hpp` are not ideal
(even if the implementation is correct): once `d` is used
as an index `[0; simDim]` and once as an index `[0; nComponents]`
and same mixed with `i`.

we now use `n` consequently for vector components (x, y, z and w)
of fields and `d` for *spatial* counts from `0` to `simDim`.
`i` is still used if we iterate over linear memory.