Refactor ionization energies for multiple stages
The ionization energies and the necessary ionization field strengths are now
stored in `PMACC_CONST_VECTOR()` arrays and can be accessed by the ionization
algorithm according to the current charge state.

Also with recent changes to the paramSets by @psychocoderHPC the cmake flag
for the gas shape has been removed.