ADIOS: Transform Empty Blocks Work-Around
Writing and reading empty/"NULL" blocks causes crashes during
reads if "zlib" transforms are enabled. We skip the read and
write of these blocks since these calls are not collective
(and `adios_define_var` is only defined temp. in memory).

Collective functions for writing: `open`, `should_buffer` and `close`.
Every other write function is local (as long as the PHDF5 transport is
not used).
For reading `init`, `open`, `perform_read`, `advance_step` and
`finalize` are collective, but `schedule_read`, `close` and the
others are not.

Thanks to Norbert Podhorszki (@pnorbert) for providing support
with minimal-examples and fast-replies to questions.