Current Interpolation: During add to EMF
This commit adds the general infrastructure to call functors
before adding `FieldJ` to the electro-magnetic fields (in
Yee: before manipulating the `E` field).

The previous method `None` without interpolation is added as a
default in `fieldSolver.param`. Additionally, a simple second order
Binomial filter for 2D and 3D is added to test the new functionality.