ADIOS Restart: Avoid NxN access pattern with Aggregators
During restart, the written N files by N aggregators are
read again by each rank which causes NxN open files and fails
for large numbers of aggregators on parallel file systems.

We now read a local block only and allgather it to calculate the
offset in the 1D partice attribute arrays again.

Due to a implementation bug in `ADIOSCountParticles.hpp:120` we
have to immitate the (wrong) Sort-by-Rank for Particles.
Particle offsets are (unfurtunately) calculated by rank but only
the scalarPos is given in the `particles_info` object.
An upcoming change in that object with `OpenPMD` `particlePatches`
will change that.