Particle Manipulators: Make Quasi-Binary
Particle manipulators are now quasi-binary, that means they
always take *two* particle species as an attribute
(+ additional parameters such as localCellIdx and 2x isParticle).

This allows to use the manipulators both during
`Manipulate` of a single species and during `CloneSpecies`.
A new `ManipulateCloneSpecies` functor was introduced that
allows for easier cloning-and-adjusting of, e.g., gas profiles
that contain of multiple ion species and need to adjust the
weighting of the created electrons directly during clone
(when cloned in a common electron species).