Fix NoSolver in 2D
With simDim equal to 2 (2D3V simulation), the MaxwellSolver
`NoSolver` did cause a compile error due to a missing trait
specialization.

This commit fixes that, in 2D3V simulations we still have 3
components of each vector field (E and B) but only 2 directions
for neighbors (the third direction is assumed to be "periodic"
or "infinitely identical").

Thanks to @n01r for reporting the issue!