Fix GaussianCloud: Powers
The implementation in `dev` for the gas profile GaussianCloud
was broken with the following effect: powers != 2 (which is the
standard Gaussian) do not permute and caused to initialize
a "smoothed cuboid" instead of an ellipsoid.

This commit fixes that by calculating r0 over sigma first,
instead of working component-wise. The previous implementation
in `master` as of the `0.1.0` release was correct, too.