Refactor FieldTmp Particle Solvers
Refactors the particle solvers that populate `FieldTmp`.

The primarily goal of this refactoring is to make the current structure
cleaner with separate classes and to allow later on to add solvers that
populate `FieldTmp`, e.g., from pure field sources (e.g., div of E, |S|, etc.).

Additionally, namings and results of solvers have been clearified and
standardized.

Changes:
 - all namings (according to openPMD ED-PIC extension, version 1.0.0)
 - <species>_chargeDensity (previously called wrong "density", related to #909)
 - <species>_energyDensity (now without charge, previously created an
                          energy-charge-density)

Adds:
 - <species>_density (an actual density)