PNGwriter 0.5.6: Bug Fix and Refactoring
Updates the requirements to use PNGwriter 0.5.6+

The latest release fixes a bug with creation time meta data that
can potentially cause data corruption (not seen in the wild with valgrind)
and refactors several parts of internal code due to coverity-scan
results.

The compile suite is already updated and a ticket for a new module
on `hypnos-hzdr` was opened.