Laser Wavepacket: Implementation != Doc
Follow-Up to #1232: the implementation of the laser wavepacket
regarding the *beam waist* does not fit its documentation in
`laserConfig.param`.

This commit updates the implementation to match the documentation
which is then identical to the other laser pulses.
Currently, the parameter that was specified is a factor `sqrt(2)`
off making the pulse waist effectively *tighter* by this factor than
intended.