Resource monitor class and resource log plugin
The resource monitor is located in libPMacc and provides
information about allocated resources of the library.
Currently, particle count and cell count per rank are
provided.

The resource log is a user adaptable resource information
log plugin which collects resource information and writes
these information to a stream. The user can choose which
resource properties to write, which format to use and
to which stream the information should be written.