openPMD: Particle Patch Reader
Implements a helper class `PatchReader` to read libSplash HDF5 files
with particle patches and to build up a class containing the full
list of particle patches for each rank.
For storing the read patches in RAM on each node, a common
`ParticlePatches` storage class was implemented (that can also be used
with the ADIOS plugin later on).

Compilation of the HDF5 class only occures when libSplash was found
(guarded by `ENABLE_HDF5` macros).