Particles: Weighting Traits
This commit adds a new trait to the particle attributes
describing each attribute's correlation in PIConGPU regarding
macro particles.

 ### MacroWeighted

Depending on the implementation of an attribute, it might be
sometimes useful to return a quantity regarding one of the underlying
real particles (false: "this attribute is not weighted accordingly
for the whole ensemble of particles in its macro particle) or just
handle the whole macro particle at once
 (true: "this attribute is already weighted").

This trait defines for each attribute if it needs to be scaled with
the weighting. *How* the scaling with weighting is applied can be
queried in the trait `WeightingPower`.

 ### WeightingPower

This trait defines how each attribute needs to be scaled with the
 weighting (linear, quadratic, ...) to convert between "real" and
"macro" particle attributes.

 ### Further references

Besides the fact that these new traits describe particle attributes
*inside* the code better and consistenly, they will also be used in
the `ED-PIC` extension of `openPMD`.

For futher information please refer to the `openPMD 1.0.0` documents:
 - https://git.io/vwlWa
 - http://www.openPMD.org
 - http://dx.doi.org/10.5281/zenodo.33624