Fix the invalid read access for fully ionized ion
When the ion is fully ionized the trait GetIonizationEnergies will not
have a valid target for read access in the vector of ionization energies.
An if condition avoiding that access has been added to the ionization model ADK.

Also equals sign alignment has been removed as it is not part of our
coding style guideline.