openPMD: ADIOS
Implements the openPMD standard including the ED-PIC extension for
our ADIOS plugin.

After first visual tests, this works with the latest ADIOS 1.9.0+
release but some string endings are broken if the latest master of
ADIOS is not used.

What is not implemeted is the refactoring of `particlePatches` for
openPMD (optional and recommended). This can be done in a follow-up
PR and also after the next release. For now this does still work and
one can generate the particle patch information in post-processing
(I have done this successfully on Titan during the last year).