Make use of faster RNG that saves & changes states
Exchange the RNG (Random Number Generator) for a faster version in
ADK_Impl.hpp (and all future ionization models). Instead of creating new
states for the RNG during each time step the state is now saved. It has the
dimensions of the grid and is mapped to the cells. Whenever a new random number
is created the state changes in that cell. Therefore it is also unique between
time steps and particle species but the RNG does not do any superfluous work.

Kudos to @FlameFire

Status: compiles [x], runs [x]