SubGrid: Set globalDomain Offset
This fixes a bug that the offset in the `PMacc::Selection`
returned by `Environment<simDim>::get().SubGrid().getGlobalDomain()`
was always set to zero (aka not set).

The implementation does not change anything else, since the
affected object was only used in a conversion from Domain to a Window
which naturally has to remove the `globalDomain.offset` (aka "set it
to zero=default") anyway.