PMacc/PIC: Update API Namings totalCellIdx
Updates the API namings and internal vars
of the copy, concat and split operations between list of
frames and linearized memory.

This does not change the logic within, so one can
still use it as an globalCellIdx converter as long
as one puts the "old" `localDomain.offset` in. The resulting
particle attribute for the domain's cellIdx can
be choosen by the user via its identifier.