CMake: Find non-default zlib
Installing PIConGPU on a minimal distribution in user
mode often includes compiling zlib yourself in a non-default
directory. This changes the way we look for zlib to find it
properly.

Notes:

zlib is not used by PIConGPU directly but comes in with a lot
of dependencies that do not declare it properly (which only
affects linking).
`m` is a similar candidate but is usually shipped with the
compiler stdlib and also has no find module in CMake afaik.