Restart: Allow Disable of Moving Window
This does not break restarts that keep the moving window "-m" on and
allows furthermore to restart with a disabled the moving window.

Warning: *enabling* the moving window from a checkpoint that
had no moving window or that was "paused" with that method will not
work. This would require remembering "how many iterations did
someone stop the moving window".

First build for Heiko during his Master's thesis and requested
again by Bifeng on the user mailing list.