openPMD: iterationFormat only Basename
The `iterationFormat` in `openPMD`'s base attributes in `/`
should only contain the *basename*, not the path to the file
in "fileBased" `iterationEncoding`.

This is not a minor bug since most post-processing tools I am aware
of are currently not reading that attribute, but they might. Nevertheless,
it violates the standard (I did not check it carefully when moving
`.h5` and `.bp` files to their own dir in `simOutput`).