ManipulateDeriveSpecies: Refactor Functors
`RatioWeightingImpl` & `ProtonTimesWeightingImpl` are without
parameters, so we need no `Impl` interface which clutters the
user interface. (Reminder: Manipulators with parameters are
first configured in `particleConfig.param` via their `Impl`
interface and then used in `speciesInitialization.param`).

`RatioWeighting` is renamed to `DensityWeighting` since it
changes the weighting depending on the `denistyRatio` of
both species.