DataConnector: Shared Pointer Storage
Rebuilds the PMacc `DataConnector` to serve as a storage to
share `shared_ptr`s with.

**Motivation**

Previously, resources in simulations that are only created once
and used from several places needed to be held by a "master" class
which keeps them "alive" in form of members. This is cumbersome and
leads in most cases to a member in `MySimulation` (which is a class
we need to reduce in complexity anyway to allow better APIs to it
and more user-centric flexibility).

Additionally, this design made it hard to call a functor which
calls a factory, e.g. for each particle species, and then
allocating new fields *without returning those objects* as done in
`CallXY` functors called by `ForEach` loops for N species
within `particles/ParticlesFunctors.hpp`. I currently need that
functionality since I want to add additional helper fields for a
user-selected subset of species without adding more book-keeping
in `MySimulation` of the `Particle`s themselves.

The new `DataConnector`, which is as before a Singleton, now serves
as an easily accessible storage that can share (e.g. as the last
owner) access to a resource which avoids creating an other singleton
for each data set with singleton-characteristics per simulation
that we have already.

**Why Not References / Pointers?**

In contrast to exchanging references as before (or plain pointers)
one can now keep track of the number of shared accesses to a
resource. The syntax of exchanging a smart-pointer is also slicker
as reference passing, which is why we did unwind the references to
(plain) pointers in most-places already.

**Other Changes**

One can now deregister selected resources and the member namings
have been cleaned up to avoid word-doubling, etc.

The sorting abstraction was removed since we only need as before
a strictly sorted FILO storage. A `vector` is preferred over a `list`
for slight performance benefits (re-allocations during sharing a
data resource initially is only done once, have less performance
penalty then one would think, otherwise we can still `reserve()`
and following frequent find-and-access is the main usage).