Ionization: Fix Charge of Ionized Ions
The charge of all but hydrogen ion species, as seen by the PIC
algorithm, was wrong after ionizing an atom/ion.

Accidentially, the charge state inside our `GetCharge` trait was
calculated as
```
    proton number * ( proton number - bound electrons)
```
instead of
```
    proton number - bound electrons
```

in case `chargeRatio` of the species in `speciesDefinition.param`
was different from `1.0` (we usually set it to the fully ionized
species' charge density).

**Affected Setups**

If all of the following criteria are true for at least one
of your ion species, your simulation is affected:

- your species in `speciesDefinition.param` used an `ionizer`
- your species is not hydrogen
- your species moves and/or creates currents (pusher, current solver)
- you defined `chargeRatio` for that species different from `1.0`

Pre-ionized setups without ionization physics are unaffected.

**Implications**

The following quantities were falsely calculated:

- plasma dynamics (pusher used wrong charge to mass ratio,
  current solver used wrong charge)
- outputs to files (`fields/<species>_chargeDensity`)

Note that the wrong plasma dynamics resulting from that highly
overcharged ion species is a severe bug and will likely render
the complete simulation unphysical if the above mentioned criterias
are met.

**Kudos**

Thanks to Thomas Kluge for reporting the issue.

Conflicts:
	src/picongpu/include/traits/attribute/GetCharge.hpp