Remove BOOST_BIND_NO_PLACEHOLDERS
Remove the general definition of `BOOST_BIND_NO_PLACEHOLDERS`
which breaks all usages below this header of `boost::placeholders`
which are in the context of `boost::bind` often assumed to be
present. This is especially true for some boost internals, which
due to that break for us from time to time.