openPMD: Fix Grid Spacing, GlobalOffset
our PIConGPU arrays describing cell sizes, offsets and sizes
are in a rather inconsequent `{x, y, z}` order while the
datasets they describe are *labeled* in `F[z][y][x]` order.

This fixes the same issue as previously seen with `axisLabels`
for openPMD in ADIOS and HDF5 output.

Only affects if at least one of the following is true:
- moving window is used
- non-quadratic cells are used

the size / layout of the simulation box (`-g`) is independent
of this issue and works quadratic and non-quadratic.