clang compatible static assert usage
An expression within a static assert those not depend on a template
can be evaluated from the compiler even if the code is never instantiated.
This means that a context check inside the body of a class is not possible with some compiler
e.g. clang.
A workaround is to extent the condition for the static assert with a template dependent
expression those is always evaluated to `true`.
The expression `sizeof(AnyType)!=0` is always `true` and can be used as workaround.