remove useless `PMACC_VERIFY`
It is not allowed to use `PMACC_VERIFY` in a device or host/device function.
  - remove `PMACC_VERIFY_MSG` from `laserTransversal`
Never the less the check is useless because it is already checked at compile time.

I added the compile time range also to the method `laserLongitudinal` to be sure that
the value is allowed before it is used inside the followed loop.