Example: Foil (LCT) TNSA
This example models a laser-ion accelerator in the TNSA regime.
An optically over-dense target n_max = 192 n_c consisting of a
liquid-crystal material is used.
Irradiated with a high-power laser pulse with a_0 = 5 the *8CB*
(4-octyl-4'-cyanobiphenyl) C_{21} H_{25} N target is assumed to be
once pre-ionized due to realistic laser contrast and pre-pulses to
C^+, H^+ and N^+ while being slightly expanded on its surfaces
(modeled as exponential density slope).

The laser is assumed to be in-focus and approximated as a plane wave
with temporally Gaussian intensity envelope of tau^{FWHM}_I = 25 fs.

This example is used to demonstrate:

* an ion acceleration setup with
* non-trivial initial conditions
* composite materials
* ionization models (compile test for ADK lin., close #1962)

with PIConGPU.

See:
    S.C. Wilks, A.B. Langdon, T.E. Cowan, M. Roth, M. Singh,
    S. Hatchett, M.H. Key, D. Pennington, A. MacKinnon, and
    R.A. Snavely.
    *Energetic proton generation in ultra-intense laser-solid*
    *interactions*,
    Physics of Plasmas **8**, 542 (2001),
    https://dx.doi.org/10.1063/1.1333697