Fix effective principal quantum number `nEff`
Fixes the calculation of the effective principal quantum number `nEff` for
the implemented ADK ionization model. The quantity `nEff` arises from scaling
an ionic potential of an arbitrary configuration and charge state Z to the
hydrogenic Bohr model. In the Bohr model:

    n = Z / sqrt(2 * E_n)

where E_n denotes the energy of the n-th level/shell.
In order to extrapolate that to arbitrary ions one has to do the following
replacements:

    Z   (nuclear charge)    --> Z_i (ionic residual charge)
    E_n (n-th shell energy) --> E_i (ionization energy of the state)

Here atomic residual charge means charge AFTER ionization since this is the
charge that the electron to be ionized 'feels'.