Rename BSIHydrogenLike to BSI and fix charge Z
The `BSIHydrogenLike` model was much too narrow since it was only applicable
to hydrogenlike states, i.e. states that have charge (Z-1)+.
The following replacement was done in

    F_BSI = E_i^2 / (4*Z)

    Z (nuclear charge) --> Z_i (ion residual charge)

That means that in a naive way the inner electrons shield the nuclear charge
from the electron that is to be ionized.

---

Also the documentation has been refined to better describe the assumptions.