Refactor ionization models to fit new routine
The new particle creation routine requires the ionization models to have
a certain structure:

    - particleCreator.numNewMacroParticles(...) returns an unsigned number of
      new macroparticles to create
    - particleCreator(...) (the functor call) creates a new particle in
      the frame slot it is given

Thus it becomes available now that the electrons that are created can be given
model-specific attribute changes.
Furthermore the behaviour of all ionization models has been matched to return
just a number of new electrons while the reduction of the bound electron number
happens only when the new electron is really created.
We are thus left with an always valid ion state between functions.