Boost Transient Deps: date_time, chrono, atomic
The Boost.threads library pulls in date_time and chrono as additional
dependencies that need a compile.

```bash
$ grep -iR "_Boost_THREAD_DEPENDENCIES" /usr/share/cmake-3.7/Modules/FindBoost.cmake
    set(_Boost_THREAD_DEPENDENCIES date_time)
[...]
    set(_Boost_THREAD_DEPENDENCIES chrono system date_time)
    set(_Boost_THREAD_DEPENDENCIES chrono system date_time atomic)
[...]
```

Since the boost *transient dependencies* are updated via a manual
list currently, we explicitly request all transient libraries for now
(instead of letting the cmake script find and add them). This helps
users with a newer boost version but an older CMake that does not
know the latest boost releases (yet).