New Tool: pic-edit
For easier forward-compatibility, we should not fill-up
add default `.param` and `.unitless` files during pic-create.

Our defaults are sane (basic PIC cycle without any density or
lasers) for a PIC simulation and will only cause unnecessary conflicts
when the defaults change.

Instead, we can clone only the files that are changed in input
on `pic-create` and add the others automatically when needed.

The workflow possible with `pic-edit` editing `.param` files is:

  pic-create $PICSRC/share/picongpu/LaserWakefield lwfa
  cd lwfa
  pic-edit grid laser isaac
  # will copy in a default isaac.param
  # will open the already existing grid.param and laser.param plus
  #   the added isaac.param inside the simulation_defines/param/
  #   directory

This also solves the issue of telling users "where to get additional
files from" when they change inputs, because they do not need to
know anymore.

`.unitless` files are not able to be opened with `pic-edit` since
they are for power-users only anyway.