PIConGPU: fix cmake compile
The PIConGPU CmakeLists.txt is separating host only and accelerator source files.
In the current implementation the host only and accelarator code is compiled twice (with host and device compiler).
The result of this bug is that compiling takes up to a factor of two longer than needed and all symboles are defined twice which can lead in linker issues.

Fix: compile `main.cpp` with the device compiler and all other files with the host compiler