Improved the cpuNumaStarter.sh script to support not using all hardware threads.
cpuNumaStarter always uses all hardware threads of a core, but depending on the
architecture this might not be what the user wants. Using more hardware threads
enables more independend register sets, but reduces the L1 cache per thread,
which might decrease performance.

The number of threads use from OpenMP is now calculated from the number of
cores per numa node and the number of hardware threads which shall be used.
If this number is not set, it is preset to the total amout of hardware threads
establishing the same behaviour as before.