Made pic-configure's back end argument independent from compiler
Although most compilers these days understands gcc's way of giving
architecture hints (-march=<architecture> -mtune=<architecture>)
this approach does not work for every compiler (like xlc or gcc
on power8 systems) and not for every case, especially of not using
"native" e.g. for crosscompiling. The Intel compiler e.g. can also
cross compile for the KNL using -xmic-avx512 (-march does not work
here). So I added a function, which checks for the used C++
compiler but also the underlying machine (for power8) to choose
the right option to pass the architecture.