CMake: Boost Target Check
It turns out, the `Boost::boost` target for the header only
libraries is always available, even if the other targets are
not, e.g. due to a newer boost version which leads to a
`_Boost_IMPORTED_TARGETS FALSE`.

This fixes the detection properly in such a case by testing for
the first (or any) of the oder requested targets instead.