PMacc CPU Architecture
Set the CPU architecture via `PMACC_CPU_ARCH`.
Defaults to an environment variable of the same name.
Only set if the (host) compiler is known, otherwise use
CXXFLAGS if you experiment with more compilers.

Only set as a convinient setter for the user instead of going
directly to CXXFLAGS as the user. Helpful in commands such as
`pic-configure` to avoid building a compiler lookup list therein.