InitController: Unphysical Particles
Allow to initialize unphysical particles in the
`InitialiserController`.
Unphysical in the sense, that those particles have mass and/or
charge *undefined*. In such a case, it makes no sense to try to
output a plasma frequency from it.

Note: we allow to define particles without such fundamental types
in PIConGPU (instead of falling back to zero) since this allows us
to differenciate between physical particles (such as photons or
neutrals) and unphysical particles (test particles, probes, etc.).