HDF5: Allow Unphysical Particle Dump
Allow to dump particles that do not have a charge or mass
frame attribute in HDF5.

Note: we allow to define particles without such fundamental
types in PIConGPU (instead of falling back to zero) since this
allows us to differentiate between physical particles (such as
photons or neutrals) and unphysical particles
(test particles, probes, etc.).

If such a particle species in encountered, we simply do not
write the constant openPMD records for charge and mass
(instead of writing implicitly a zero value, which would
suggest these particles are physical - which they are not).