PluginController: Eligible Species
Only specialize and compile particle plugins for particle species
which are elibible for it, based on the `SpeciesEligibleForSolver`
trait.

Silently disables components which are not fulfilled, e.g. for
nonphysical particles such as probes. Moves user errors from
compile time (which are in this case always very long and unreadable
template meta programming errors) to runtime where it is (ideally)
easier to spot.

Possible follow-up: For further usability, we can add log messages
on high-level debug levels to output which species are ignored on
`pluginLoad()`.