Add lower temperature cutoff for Thomas-Fermi
Depending on the material the Thomas-Fermi prediction for the average charge
state can be unphysically high. For some materials it predicts non-zero
charge states at 0 temperature. That can be a reasonable approximation
for metals and their electrons in the conduction band.
In general there should be no free electrons at 0 temperature, though.

Therefore this commit implements a lower temperature cutoff.
The `detailedBalanceThomasFermi` function now accepts the parameters

    * `temperature` in eV
    * `massDensity` in g/cm^3

just as the fitting model described by More (1985).