In-Source Builds: CMake is too accurate
The check should prevent in-source-builds.
If your source directory is picongpu and the build directory is
`picongpu_build`, this check throws an error. But this isn't an
error. To prevent this, I add a /. This should be correct because
`SOURCE_DIR_ROOT` is a file (It is the result of
get_filename_component(...) ). I checked it and get no error with
the directory names picongpu and picongpu_build. The directory
picongpu/build throws an error message.