(Manipulate)DeriveSpecies: Fix Filters
The filters on `ManipulateDerive`- and `Derive`-Species should be
filtering the particles to derive (in the source species). Before,
this was accidently filtering the destination species to manipulate
and derived all particles from the source species. This is not
necessary, lacks the filter in the source species and is confusing,
also since a manipulater can simply filter with a simple if where
needed.