Docs: Hide Double ToC
In the HTML version of our docs, the table of contents and the
general titles was unnecessarily shown.

It is accessible in the left side bar, so the ToC can be hidden.
Still, the manual chapters are needed for proper pdf chapters,
which does grouping by ToC only while ToC captions are
irrelevant.

This keeps the old structure which works on both formats (and epub)
but hides the superfluous titles via css.