Param Docs: Title Only
In the section that introduces the usage of .param files,
only the titles of the files shall be shown. This is currently
not possible, see https://github.com/michaeljones/breathe/issues/318

This PR applies a work-around for HTML to hide additional doxygen
information (does not fix the underlying issue and is issue still
exists in pdf & epub).