Hypnos: Zlib
Hypnos did not yet properly build zlib in the dependency chain.

This led to several severe instabilities and bug reports in the past
months, mainly with unpredictible crashes & segfaults during I/O but
also with compile errors in CPUs.

While modules are getting rebuild, a lack of auto-loading the
dependency means we have to load zlib manually for now.