optimize stride size for update FieldJ
The stride size for `StrideMapping` depends on the supercell size and the stencil margins.
In the case where the lower and upper margin fits into one supercell we need only a stride of 2 instead of the current used default stride 3.
If we use stencils with margins larger than one supercell the default 3 can be to low and will produce data race conditions (therefore this PR is also a fix for big stencil margins).

- calculate optimal stride size (skip only as much supercells as needed to avoid race conditions for cell updates)