fix that last frame can be empty
By definition our last frame is allowed to not full filled with particles.
To reduce the memory foot rpint the last frame should never be empty
after all gaps are filled.
This is currently not always the case.

Remove the last frame in the a supercell in `KernelFillGapsLastFrame` if it is empty.