Remove unnecessary m_species
In `PhaseSpace.tpp` both the `axis_p_range` as well as the `pRange_unit`
contain the factor m_species (species mass in electron masses).
That resulted in the momentum range being too large for all ions
when later momentum values are multiplied by this `pRange_unit`.
This commit fixes that issue.