fix tbg error message
If the path to `egetopt` is not in the environment variable `PATH` tbg falls back to use the `egetopt` in the same location as `tbg` itself. In that case the command `which` is throwing a long error message that egetopt can not be found.
This PR hide the error message and create an own error message if needed.