fix compile with boost 1.68.0
Boost 1.68.0 is not allowed since #2685 but after we will integrate the next alapak release we will end in compile issues.
The reason is #2707. The definition of `boost::math::gcd` is not available if `boost/integer/common_factor_rt.hpp` is included.

This PR fixes the behaviour be using always `boost/integer/common_factor_rt.hpp` and use `boost::integer::gcd` within PMacc.