cuda_memtest usage on hypnos
`cuda_memtest` can currently only be used if the node is exclusive allocated by the user and no other job is running on it.
All hypnos template files for hypnos supports that the user can allocate only a subset of a node. In this case `cuda_memtest` will crash if an other user/job is already using gpu zero.
This PR will disable the GPU memory test if the job has not alloacted all GPUs.

- remove PBS option `-n` (exclusive node usage) because it is not working on hypnos
- disable the test `cuda_memtest` if node is not exclusive allocated