Docker: More Examples, NGC & Benchmark
OpenMPI: Use 2.1.2

Multi-Rank MPI support and thus multi-GPU support requires for
in-node communication a BTL such as "sm" which was replaced with
"vader" in OpenMPI 3.0.0.

"vader" requires CMA support from the kernel which seems to be
something that is not working in a docker container. We therefore
switch back to an older release (pre-3.0) of OpenMPI that still has
the (slower) "sm" transport.

Build for SM_70 (V100)

Docker Readme: MPI BTL shared Mem

Overwrite exising output with tbg -f

Configurable output directory and print to user.

Use non-ISAAC examples for now.
Add _live LWFA examples for later testing of ISAAC.