new tool: tbg-submit
Provide a tool to re submit a job created with tbg.
tbg is abstracting the batch system from the user.
In the case the user is not using `-s` during the tbg usage or needs to be restart a job he/she needs a deep knowledge how to use the batch system.
Due to the reason we have configured all for the user in our shipped profile files we need
only a tool which knows how to restart a job.