fix particle creation if density zero
fix #2823

In the case where the density functor is returning a density of zero or negative the functor which calculate the number of macro partices (called `startPosition` functor) should not be called.
This avoid that the functor write e.g. the user who is using a free functor implementation must handle those edge cases.

changes:
 - do not call `startPosition` functor if density is `<= 0.0`