CUDA 9.0+, GCC 5.1+, Boost 1.65.1+
Increase the dependencies to CUDA 9.0+, which in turn
requires Boost 1.65.1+ to run.

With that prepare, we also update our CI to run on GCC 5.1+
which will make the transition to C++14 possible soon. While
GCC 4.9 is technically still possible, we deprecate support
and continous CI for it to move forward for the aforementioned
reasons.