Dockerfile: No SSH Deamon and Keys
For local startup, an SSH daemon is not necessary.

This is a security fix, removing pre-set keys installed alongside
the openssh-server package in /etc/ssh which would be known and
published in the docker image. The published dockerhub images
ax3l/picongpu:0.4.3 and ax3l/picongpu:latest have been replaced
with a fixed version.

Affected are users that would run the docker container on a
non-private network or shared system which could lead to the
security implication of unauthorized login into the running docker
instance, if the docker SSH server port is published (note that
this is not the default and has to be exposed manually).

Thank you to Adam Simpson (Nvidia) for reporting the issue!